Order Update

+++ LaylaERP/js/Orders/NewOrder.js
@@ -693,7 +693,7 @@ function getOrderItemList(oid) {
 
                     if (coupon_list[j].type == 'add_coupon') {
                         couponHtml += '$<span id="cou_discamt">' + cou_amt.toFixed(2) + '</span>';
-                        couponHtml += '<button type="button" class="btn btn-box-tool pull-right billinfo" onclick="removeCouponInList(\'' + coupon_list[j].post_title.toString().toLowerCase() + '\');"> <i class="fa fa-times"></i>';
+                        couponHtml += '<button type="button" class="btn btn-box-tool pull-right billinfo" onclick="deleteAllCoupons(\'' + coupon_list[j].post_title.toString().toLowerCase() + '\');"> <i class="fa fa-times"></i>';
                         couponHtml += '</button>';
                     }
                     else {
@@ -712,7 +712,7 @@ function getOrderItemList(oid) {
                     couponHtml += '<span>' + cpn_name + '</span>';
                     couponHtml += '<div class="pull-right">';
                     couponHtml += '$<span id="cou_discamt">' + cou_amt.toFixed(2) + '</span>';
-                    couponHtml += '<button type="button" class="btn btn-box-tool pull-right billinfo" onclick="removeCouponInList(\'' + data[i].product_name + '\');"><i class="fa fa-times"></i></button>'
+                    couponHtml += '<button type="button" class="btn btn-box-tool pull-right billinfo" onclick="deleteAllCoupons(\'' + data[i].product_name + '\');"><i class="fa fa-times"></i></button>'
                     couponHtml += '</div>';
                     couponHtml += '</a>';
                     couponHtml += '</li>';
@@ -913,7 +913,7 @@ function Coupon_get_discount_amount(id, parent_id, coupon_code, coupon_amt, item
     if (coupon_isedu.includes(coupon_isedu)) { isedu = 1; }
     let isgrin = 0;
     if (coupon_isgrin.includes(coupon_isedu)) { isgrin = 1; }
-    
+
     if (coupon_code.includes("friend") && coupon_code.substr(6) > 8500) {
         if (id != 632713 && id != 78676) {
             if (parent_id == 118) {
@@ -1010,6 +1010,22 @@ function Coupon_get_discount_amount(id, parent_id, coupon_code, coupon_amt, item
     else {
         return { price: reg_price, disc_amt: coupon_amt, qty: item_qty };
     }
+}
+function check_applied_coupon(coupon_code, product_ids, exclude_product_ids) {
+    let check = false, rq_prd_ids = [], exclude_ids = [];
+    if (product_ids != "" && product_ids != null) {
+        rq_prd_ids = product_ids.split(",").map((el) => parseInt(el));
+    }
+    if (exclude_product_ids != "" && exclude_product_ids != null) {
+        exclude_ids = exclude_product_ids.split(",").map((el) => parseInt(el));
+    }
+    $("#order_line_items > tr.paid_item").each(function (index, row) {
+        let pid = $(row).data('pid'), vid = $(row).data('vid');
+        if (!exclude_ids.includes(pid) && !exclude_ids.includes(vid) && ((rq_prd_ids.includes(pid) || rq_prd_ids.includes(vid)) || rq_prd_ids == 0)) {
+            check = true;
+        }
+    });
+    return check;
 }
 function getAllCoupons() {
     var coupons = [];
@@ -1109,13 +1125,16 @@ function ApplyCoupon() {
             if (data[0].date_expires != "" && data[0].date_expires != null) {
                 let exp_date = new Date(data[0].date_expires);
                 let today = new Date();
-                console.log(data[0].date_expires, exp_date, today);
                 if (exp_date < today) {
                     swal('Alert!', 'Coupon code has been expired.', "info").then((result) => { $('#txt_Coupon').focus(); return false; }); return false;
                 }
             }
             data[0].coupon_amount = parseFloat(data[0].coupon_amount) || 0.00;
             data[0].limit_x_items = parseInt(data[0].limit_x_items) || 0;
+
+            if (!check_applied_coupon(coupon_code, data[0].product_ids, data[0].exclude_product_ids)) {
+                swal('Alert!', 'Can not add ' + coupon_code, "info").then((result) => { $('#txt_Coupon').focus(); return false; }); return false;
+            }
 
             if (coupon_code.includes("friend") && coupon_code.substr(6) > 8500) { deleteAllCoupons('friend_diff'); }
             else if (coupon_code.includes("freeprotector")) { }
@@ -1159,7 +1178,7 @@ function bindCouponList(data) {
 
                 if (data[0].type == 'add_coupon') {
                     layoutHtml += '$<span id="cou_discamt">' + cou_amt.toFixed(2) + '</span>';
-                    layoutHtml += '<button type="button" class="btn btn-box-tool pull-right" onclick="removeCouponInList(\'' + data[i].post_title.toString().toLowerCase() + '\');">';
+                    layoutHtml += '<button type="button" class="btn btn-box-tool pull-right" onclick="deleteAllCoupons(\'' + data[i].post_title.toString().toLowerCase() + '\');">';
                     layoutHtml += '<i class="fa fa-times"></i>';
                     layoutHtml += '</button>';
                 }
@@ -1304,9 +1323,6 @@ function deleteAllCoupons(coupon_type) {
             });
     }
 }
-function removeCouponInList(code) {
-    deleteAllCoupons(code);
-}
 function freeQtyUpdate() {
     $("#order_line_items > tr.free_item").each(function (index, row) {
         let zQty = 0.00, pid = parseInt($(this).data("pid")) || 0;
@@ -1344,7 +1360,7 @@ function calculateDiscountAcount() {
     });
     //Calculate discount
     $('#billCoupon li.items').each(function (index, li) {
-        let cou_amt = 0.00, cou = $(li).data('coupon').toString();
+        let cou_amt = 0.00, cou = $(li).data('coupon').toString().toLowerCase();
         let zCouponAmt = parseFloat($(li).data('couponamt')) || 0.00, zDiscType = $(li).data('disctype'), zType = $(li).data('type'), zQty = 0.00, zRegPrice = 0.00, zSalePrice = 0.00, zGrossAmount = 0.00, zDisAmt = 0.00;
 
         let rq_prd_ids = [], exclude_ids = [];
@@ -1399,6 +1415,7 @@ function calculateDiscountAcount() {
         $(li).find("#cou_discamt").text(cou_amt.toFixed(2))
         if (zDiscType == '2x_percent' && cou_amt > 0) $(li).removeClass('hidden');
         else if (zDiscType == '2x_percent') $(li).addClass('hidden');
+        //if (cou_amt == 0) deleteAllCoupons(cou);
     });
     calcFinalTotals();
 }
